Fixed Satin Makefiles: make test. Added some options to the Ibis Latency test. You can now say where to finish the messages when upcalls are used, or not to finish them at all. Renamed Manta Streams -> Ibis Streams. Added a new layer: SerializationStreams. This is a generic serialization interface, including support for array slices, arrays of primitive types and objetc replacement. Below this are *currently* two implementations: Sun and Ibis serialization. Because the global interface is now the same, the Ibis implementations don't have to write seperate read and write messages, etc. for all serialization types. This way, the size of the TCP implementation was reduced by 35%.
Made the interface of the read and write messages cleaner and more consistent.
Even added some Javadoc comments :-)

Added support for firewalls in the TCP implementation.
Use the property -Dfirewall_range=3000-4000 to tell Ibis that the port range
between 3000 and 4000 is open, and should be used by Ibis. All ports will then
be allocated within this range.

Updated the messagePassing impl to support the new streaming layers.

Fixed several nameserver bugs and race conditions, especially when nodes leave.

Rewrote a large part of the TCP impl. The upcall/finish behaviour is now
correct. A new thread is started to handle upcalls when a message is finished.
Also, fixed some race conditions where two threads could simultaneously read
from the same socket.

Removed RMIOutputstream, the replacement mechamisn is now the same for Sun and
Ibis serialization, so this is not needed anymore.