Lots of tiny tweaks and changes to make the system visualization behave!
Re-enabled the "skip lowest level" option as default in the gathering tweaks, because I now remember what it was there for in the first place (to prevent a location being made for each single ibis, and actually allow locations to be meaningful). 
All of the visual-tweak sliders will now work according to the original design again.